Fixed a missing places problem (e.g., 7.01.59.02).
The S&G database has some of the places duplicated on ptolemy id
to show an alternate name. For some (all?) of these, the ptolemy
coordinate set is missing. The previous code dropped duplicates
on ptolemy id first, and then removed ones without a ptolemy
coordinate set. This caused some locations (such as 7.01.59.02)
to be completely dropped from further processing. Reversing the
order of the operations (filter on missing ptolemy coordinates,
then remove duplicates on ptolemy id) fixes the problem.