Implemented features needed to process Arabia.
Dmitri sent a file for Arabia Felix as a spreadsheet, formatted
the way the previous output spreadsheets had been formatted. To be
able to read from this instead of the S&G database as before, a
new read function was needed, along with its dependency on the xlrd
package, as well as changes to the main driver program to specify
some arguments. I also took the opportunity to allow additional
arguments to specify the id prefix in the case you are using sgdb
(previously it was hardcoded and I was manually changing between 7.01
and 7.04 to go from India mainland to Sri Lanka for instance), as well
as change the name of the output file.