Run multiple models on multiple inputs.
We're getting enough datasets now and have enough models
and enough frequent updates that it probably makes sense to
start automating the pipeline a bit. This is a first step,
a simple script to run each of the main current models,
triangulate and flocking, on each of the four inputs we're
working with for Arabia.

Still to do is removing the local path references and replacing
them with proper command line arguments, and adding in the
India regions, and perhaps looking at a better workflow manager
such as luigi, which I've used on other projects and works well.