Correct the weights.
Dmitri discovered while re-reading our Corfu 2015 paper that
the weights were being applied to the flocking algorithm such
that points that were further away from the unknown point in
Ptolemy space were being weighted higher than points that were
closer. This fix reverses that, so that points that are closer
are weighted higher, according to an improved formula that
Dmitri provided.