Make bayesian adj. to work with other regions.
The bayesian adjustment logic was hard coded to India. Added some
basic parameters to get it to work with Arabia, and added a simple
batch file to adjust all the Arabia files we have. Also added a
script to generate an image that can be used to create the prior.